Move PerResourcePositionsService to r2-shared
