+++ r2-streamer/src/main/java/org/readium/r2/streamer/parser/divina/DiViNaParser.kt
@@ -13,10 +13,9 @@ import org.json.JSONObject
 import org.readium.r2.shared.format.MediaType
 import org.readium.r2.shared.publication.Manifest
 import org.readium.r2.shared.publication.Publication
-import org.readium.r2.shared.publication.services.positionsServiceFactory
+import org.readium.r2.shared.publication.services.PerResourcePositionsService
 import org.readium.r2.streamer.BuildConfig.DEBUG
 import org.readium.r2.streamer.container.ContainerError
-import org.readium.r2.streamer.parser.PerResourcePositionsService
 import org.readium.r2.streamer.parser.PubBox
 import org.readium.r2.streamer.parser.PublicationParser
 import timber.log.Timber
