+++ r2-streamer/src/test/java/org/readium/r2/streamer/parser/epub/EpubDeobfuscatorTest.kt
@@ -33,7 +33,7 @@ class EpubDeobfuscatorTest {
         val deobfuscationDir = EpubDeobfuscatorTest::class.java
             .getResource("deobfuscation/cut-cut.woff")
             ?.path
-            ?.let { File(it).parent }
+            ?.let { File(it).parentFile }
         assertNotNull(deobfuscationDir)
         fetcher = FileFetcher("/deobfuscation", deobfuscationDir)
